+++ actionmailer/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Add structured events for Action Mailer.
+
+    *Gannon McGibbon*
+
 ## Rails 8.1.0.beta1 (September 04, 2025) ##
 
 *   Add `deliver_all_later` to enqueue multiple emails at once.